project page resource page new toolbar intergration and actuals page toolbar removed
