+++ app/components/Resources/ResourceTable.tsx
@@ -13,7 +13,6 @@ import {
 import ResourceToolbar from '../Toolbar/ResourceToolbar';
 import { JSXElementConstructor, useState } from 'react';
 import { Resource } from '@/app/types';
-import CommonToolbar from '../Toolbar/CommonToolbar';
 import { Box } from '@mui/material';
 
 interface ResourceTableProps {
@@ -48,7 +47,6 @@ const ResourceTable = ({
   const [filterButtonEl, setFilterButtonEl] = useState(null);
   return (
        <Box sx={{ display: 'flex', flexDirection: 'column',height:'95vh' }}>
-        <CommonToolbar />
     <StyledDataGrid
       apiRef={apiRef}
       columns={columns}
