+++ app/components/Resources/TeamsTable.tsx
@@ -13,7 +13,6 @@ import {
 import ResourceToolbar from '../Toolbar/ResourceToolbar';
 import { JSXElementConstructor, useState } from 'react';
 import { Team } from '@/app/types';
-import CommonToolbar from '../Toolbar/CommonToolbar';
 import { Box } from '@mui/material';
 
 interface TeamsTableProps {
@@ -48,93 +47,90 @@ const TeamsTable = ({
   const [filterButtonEl, setFilterButtonEl] = useState(null);
 
   return (
-     <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
-            <CommonToolbar />
-    <StyledDataGrid
-      apiRef={apiRef}
-      columns={columns}
-      rows={rows}
-      hideFooter={true}
-      hideFooterSelectedRowCount={true}
-      loading={loading}
-      initialState={{
-        sorting: {
-          sortModel: [
-            { field: 'Team', sort: 'asc' },
-          ],
-        },
-        columns: {
-          columnVisibilityModel: {
-            Team: true,
+    <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
+      <StyledDataGrid
+        apiRef={apiRef}
+        columns={columns}
+        rows={rows}
+        hideFooter={true}
+        hideFooterSelectedRowCount={true}
+        loading={loading}
+        initialState={{
+          sorting: {
+            sortModel: [{ field: 'Team', sort: 'asc' }],
           },
-        },
-      }}
-      slots={{
-        toolbar:
-          ResourceToolbar as unknown as JSXElementConstructor<GridToolbarProps>,
-        columnMenu: CustomColumnMenu,
-      }}
-      localeText={{
-        toolbarFilters: '',
-        toolbarColumns: '',
-      }}
-      sx={{
-        height: '95vh',
-        '& .MuiDataGrid-columnHeader': {
-          padding: '0 46px',
-          borderRight: 'none',
-        },
-        '& .MuiDataGrid-footer': {
-          display: 'block',
-        },
-      }}
-      slotProps={{
-        loadingOverlay: {
-          variant: 'skeleton',
-          noRowsVariant: 'skeleton',
-        },
-        panel: {
-          anchorEl: filterButtonEl,
-          className: 'parent-grid-panel',
-        },
-        toolbar: {
-          //@ts-ignore
-          setFilterButtonEl,
-          value: value,
-          onChange: onChange,
-        },
-        columnsPanel: {
-          className: 'styleColumnMenu',
-          sx: ColumnManagementStyles,
-        },
-        filterPanel: {
-          columnsSort: 'asc',
-          //@ts-ignore
-          className: 'filterPopup',
-          filterFormProps: {
-            columnInputProps: {
-              size: 'small',
-              sx: { mt: 'auto' },
+          columns: {
+            columnVisibilityModel: {
+              Team: true,
             },
-            operatorInputProps: {
-              size: 'small',
-              sx: { mt: 'auto' },
-            },
-            valueInputProps: {
-              InputComponentProps: {
+          },
+        }}
+        slots={{
+          toolbar:
+            ResourceToolbar as unknown as JSXElementConstructor<GridToolbarProps>,
+          columnMenu: CustomColumnMenu,
+        }}
+        localeText={{
+          toolbarFilters: '',
+          toolbarColumns: '',
+        }}
+        sx={{
+          height: '95vh',
+          '& .MuiDataGrid-columnHeader': {
+            padding: '0 46px',
+            borderRight: 'none',
+          },
+          '& .MuiDataGrid-footer': {
+            display: 'block',
+          },
+        }}
+        slotProps={{
+          loadingOverlay: {
+            variant: 'skeleton',
+            noRowsVariant: 'skeleton',
+          },
+          panel: {
+            anchorEl: filterButtonEl,
+            className: 'parent-grid-panel',
+          },
+          toolbar: {
+            //@ts-ignore
+            setFilterButtonEl,
+            value: value,
+            onChange: onChange,
+          },
+          columnsPanel: {
+            className: 'styleColumnMenu',
+            sx: ColumnManagementStyles,
+          },
+          filterPanel: {
+            columnsSort: 'asc',
+            //@ts-ignore
+            className: 'filterPopup',
+            filterFormProps: {
+              columnInputProps: {
                 size: 'small',
+                sx: { mt: 'auto' },
+              },
+              operatorInputProps: {
+                size: 'small',
+                sx: { mt: 'auto' },
+              },
+              valueInputProps: {
+                InputComponentProps: {
+                  size: 'small',
+                },
+              },
+              deleteIconProps: {
+                sx: {
+                  '& .MuiSvgIcon-root': { color: '#d32f2f' },
+                },
               },
             },
-            deleteIconProps: {
-              sx: {
-                '& .MuiSvgIcon-root': { color: '#d32f2f' },
-              },
-            },
+            sx: FilterPanelStyles,
           },
-          sx: FilterPanelStyles,
-        },
-      }}
-    />
+        }}
+      />
     </Box>
   );
 };
